+++ mapbox-maps-ios/Sources/MapboxMaps/Location/LocationOptions.swift
@@ -25,8 +25,8 @@ public struct LocationOptions: Equatable {
 
     /// Specifies if a `Puck` should use `Heading` or `Course` for the bearing.
     ///
-    /// The default value is `PuckBearingSource.heading`.
-    public var puckBearingSource: PuckBearingSource
+    /// The default value is `PuckBearing.heading`.
+    public var puckBearing: PuckBearing
 
     /// Whether the puck rotates to track the bearing source.
     ///
@@ -40,27 +40,27 @@ public struct LocationOptions: Equatable {
     ///   - desiredAccuracy: Specifies the accuracy of the location data.
     ///   - activityType: Sets the type of user activity associated with the location.
     ///   - puckType: Sets the type of puck that should be used.
-    ///   - puckBearingSource: Specifies if a `Puck` should use `Heading` or `Course` for the bearing.
+    ///   - puckBearing: Specifies if a `Puck` should use `Heading` or `Course` for the bearing.
     ///   - puckBearingEnabled: Whether the puck rotates to track the bearing source.
     public init(
         distanceFilter: CLLocationDistance = kCLDistanceFilterNone,
         desiredAccuracy: CLLocationAccuracy = kCLLocationAccuracyBest,
         activityType: CLActivityType = .other,
         puckType: PuckType? = nil,
-        puckBearingSource: PuckBearingSource = .heading,
+        puckBearing: PuckBearing = .heading,
         puckBearingEnabled: Bool = true
     ) {
         self.distanceFilter = distanceFilter
         self.desiredAccuracy = desiredAccuracy
         self.activityType = activityType
         self.puckType = puckType
-        self.puckBearingSource = puckBearingSource
+        self.puckBearing = puckBearing
         self.puckBearingEnabled = puckBearingEnabled
     }
 }
 
 /// Controls how the puck is oriented
-public enum PuckBearingSource: Equatable {
+public enum PuckBearing: Equatable {
     /// The puck should set its bearing using `heading: CLHeading`. Bearing will mimic user's
     /// spatial orientation.
     case heading
@@ -68,4 +68,38 @@ public enum PuckBearingSource: Equatable {
     /// The puck should set its bearing using `course: CLLocationDirection`. Bearing will mimic
     /// the general direction of travel.
     case course
+}
+
+/// :nodoc:
+@available(*, deprecated, renamed: "PuckBearing")
+public typealias PuckBearingSource = PuckBearing
+
+// MARK: Deprecated
+
+extension LocationOptions {
+
+    /// :nodoc:
+    @available(*, deprecated, renamed: "puckBearing")
+    public var puckBearingSource: PuckBearingSource {
+        get { puckBearing }
+        set { puckBearing = newValue }
+    }
+
+    /// :nodoc:
+    @available(*, deprecated, renamed: "LocationOptions.init(distanceFilter:desiredAccuracy:activityType:puckType:puckBearing:puckBearingEnabled:)")
+    public init(
+        distanceFilter: CLLocationDistance = kCLDistanceFilterNone,
+        desiredAccuracy: CLLocationAccuracy = kCLLocationAccuracyBest,
+        activityType: CLActivityType = .other,
+        puckType: PuckType? = nil,
+        puckBearingSource: PuckBearingSource,
+        puckBearingEnabled: Bool = true
+    ) {
+        self.distanceFilter = distanceFilter
+        self.desiredAccuracy = desiredAccuracy
+        self.activityType = activityType
+        self.puckType = puckType
+        self.puckBearing = puckBearingSource
+        self.puckBearingEnabled = puckBearingEnabled
+    }
 }